requirements.txt
discord.py==2.2.2
pytz==2022.7.1
